Refactored the lexer
It now reads the entire source as a string since we gain nothing in
having it as a file stream since memory usage is of less concern
for where the lexer will be used.

It now also works on characters instead of string slices.

This spead thing up something like 7-8 times.